# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для сокращения URL-адресов",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/links": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Создать короткую ссылку",
                "parameters": [
                    {
                        "description": "Данные ссылки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.CreateLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.LinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/links/{short_code}/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает аналитику кликов по короткой ссылке",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Получить статистику кликов",
                "parameters": [
                    {
                        "type": "string",
                        "example": "test123",
                        "description": "Короткий код ссылки",
                        "name": "short_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.AnalyticsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Вход в систему с email и паролем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "Создает нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/url-short_internal_models.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "url-short_internal_models.AnalyticsResponse": {
            "type": "object",
            "properties": {
                "clicks": {
                    "description": "Список кликов",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/url-short_internal_models.ClickStatistic"
                    }
                },
                "total_clicks": {
                    "description": "Общее количество кликов\nexample: 42",
                    "type": "integer"
                }
            }
        },
        "url-short_internal_models.ClickStatistic": {
            "type": "object",
            "properties": {
                "browser": {
                    "description": "Браузер пользователя\nexample: Chrome 115",
                    "type": "string"
                },
                "clicked_at": {
                    "description": "Время клика\nexample: 2024-02-20T15:04:05Z",
                    "type": "string"
                },
                "device_type": {
                    "description": "Тип устройства\nexample: mobile",
                    "type": "string"
                },
                "ip_address": {
                    "description": "IP-адрес клиента\nexample: 192.168.1.1",
                    "type": "string"
                },
                "location": {
                    "description": "Геолокация клиента\nexample: Moscow, Russia",
                    "type": "string"
                },
                "os": {
                    "description": "Операционная система\nexample: Android 13",
                    "type": "string"
                }
            }
        },
        "url-short_internal_models.CreateLinkRequest": {
            "type": "object",
            "required": [
                "original_url"
            ],
            "properties": {
                "custom_code": {
                    "type": "string",
                    "example": "my_custom_code"
                },
                "original_url": {
                    "type": "string",
                    "example": "https://google.com"
                }
            }
        },
        "url-short_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Описание ошибки\nexample: Неверные данные",
                    "type": "string"
                }
            }
        },
        "url-short_internal_models.LinkResponse": {
            "type": "object",
            "properties": {
                "full_url": {
                    "type": "string",
                    "example": "http://localhost:8080/a1b2c3"
                },
                "short_code": {
                    "type": "string",
                    "example": "a1b2c3"
                }
            }
        },
        "url-short_internal_models.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "qwerty123"
                }
            }
        },
        "url-short_internal_models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGci..."
                }
            }
        },
        "url-short_internal_models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "qwerty123"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "url-short_internal_models.RegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Пользователь создан"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}